Refactor code to re-use constants in headers
