+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardResponseParser.kt
@@ -9,6 +9,7 @@ import com.paypal.android.corepayments.OrderErrorDetail
 import com.paypal.android.corepayments.OrderStatus
 import com.paypal.android.corepayments.PayPalSDKError
 import com.paypal.android.corepayments.PaymentsJSON
+import com.paypal.android.corepayments.common.Headers
 import org.json.JSONException
 
 internal class CardResponseParser {
@@ -31,7 +32,7 @@ internal class CardResponseParser {
                 json.optMapObjectArray("purchase_units") { PurchaseUnit(it) }
             )
         } catch (ignored: JSONException) {
-            val correlationId = httpResponse.headers["Paypal-Debug-Id"]
+            val correlationId = httpResponse.headers[Headers.PAYPAL_DEBUG_ID]
             val error = APIClientError.dataParsingError(correlationId)
             ConfirmPaymentSourceResult.Failure(error)
         }
@@ -42,7 +43,7 @@ internal class CardResponseParser {
             result = null
         } else {
 
-            val correlationId = httpResponse.headers["Paypal-Debug-Id"]
+            val correlationId = httpResponse.headers[Headers.PAYPAL_DEBUG_ID]
             val bodyResponse = httpResponse.body
             if (bodyResponse.isNullOrBlank()) {
                 result = APIClientError.noResponseData(correlationId)
