+++ CorePayments/src/main/java/com/paypal/android/corepayments/common/Headers.kt
@@ -1,0 +1,7 @@
+package com.paypal.android.corepayments.common
+
+object Headers {
+    const val PAYPAL_DEBUG_ID = "paypal-debug-id"
+    const val AUTHORIZATION = "Authorization"
+    const val CONTENT_TYPE = "Content-Type"
+}
